requirements.txt
matplotlib         3.1.1
numpy              1.17.2
PySide2            5.13.1
pytest             5.2.1
PyYAML             5.1.2
